UI調節と最適化

+++ lib/view/home_page.dart
@@ -22,13 +22,13 @@ class _HomePageState extends State<HomePage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      drawerEdgeDragWidth: 0.0,
       appBar: AppBar(
+        automaticallyImplyLeading: false,
         elevation: 0,
         shape: const RoundedRectangleBorder(
           borderRadius: BorderRadius.only(
             bottomLeft: Radius.circular(25),
-            bottomRight: Radius.circular(25) ,
+            bottomRight: Radius.circular(25),
           ),
         ),
         title: const Text(
@@ -70,11 +70,12 @@ class _HomePageState extends State<HomePage> {
                               descending: false,
                             )
                             .snapshots(),
-                        builder: (context, AsyncSnapshot<QuerySnapshot> snapshot) {
+                        builder:
+                            (context, AsyncSnapshot<QuerySnapshot> snapshot) {
                           if (!snapshot.hasData) {
                             return const CircularProgressIndicator();
-                          }
-                          return ListView( 
+                          } 
+                          return ListView(
                             children: snapshot.data!.docs.map((campes) {
                               if (campes['content'] == null) {
                                 campes.reference.update({
@@ -93,15 +94,39 @@ class _HomePageState extends State<HomePage> {
                                   ),
                                   child: ListTile(
                                     dense: false,
-                                    leading: const Text('index'),
                                     title: Text(campes['content'].toString()),
                                     onLongPress: () {
-                                      campes.reference.delete();
+                                      // ダイアログはマテリアルで統一した方が美しい
+                                      showDialog(
+                                          context: context,
+                                          builder: (_) {
+                                            return CupertinoAlertDialog(
+                                              title: const Text('削除しますか？'),
+                                              content:
+                                                  const Text('削除すると元に戻せなくなります'),
+                                              actions: [
+                                                CupertinoDialogAction(
+                                                  child: const Text('キャンセル'),
+                                                  isDestructiveAction: false,
+                                                  onPressed: () =>
+                                                      Navigator.pop(context),
+                                                ),
+                                                CupertinoDialogAction(
+                                                  child: const Text('削除'),
+                                                  isDestructiveAction: true,
+                                                  onPressed: () {
+                                                    campes.reference.delete();
+                                                    Navigator.pop(context);
+                                                  },
+                                                ),
+                                              ],
+                                            );
+                                          });
                                     },
                                     onTap: () async {
-                                      var updatedContent =
-                                          await Navigator.push(context,
-                                              MaterialPageRoute(builder: (context) {
+                                      var updatedContent = await Navigator.push(
+                                          context,
+                                          MaterialPageRoute(builder: (context) {
                                         return UpdatePage(
                                             oldText: campes['content']);
                                       }));
